[clean] Refactor codes
ファイル分割、CMake用関数作成
